+++ CHANGES.rst
@@ -16,6 +16,8 @@
   (`Issue #27 <https://github.com/graingert/redshift_sqlalchemy/pull/27>`_)
 - Fix unicode issue with SORTKEY on python 2.
   (`Issue #34 <https://github.com/graingert/redshift_sqlalchemy/pull/34>`_)
+- Add support for Redshift ``DELETE`` statements that refer other tables in the ``WHERE`` clause
+  clause. (`Issue #26 <https://github.com/graingert/redshift_sqlalchemy/pull/26>`_)
 
 
 0.1.2 (2015-08-11)